Finalizado Primeira Parte da Atividade

+++ LaboratorioPadroes/src/com/laboratorio/quintaquestao/CalculadoraDeDescontos.java
@@ -1,8 +1,10 @@
 package com.laboratorio.quintaquestao;
 
+import java.math.BigDecimal;
+
 public class CalculadoraDeDescontos {
 
-	public double calcularDesconto(Cliente cliente, NotaFiscal nota){
+	public BigDecimal calcularDesconto(Cliente cliente, NotaFiscal nota){
 		DescontoStrategy regraDeDesconto = cliente.getTipo().getRegraDeDesconto();
 		return regraDeDesconto.desconto(cliente, nota);
 	}
