+++ LaboratorioPadroes/src/com/laboratorio/quintaquestao/RegraDeTaxaParaDesconto.java
@@ -1,7 +1,9 @@
 package com.laboratorio.quintaquestao;
 
+import java.math.BigDecimal;
+
 public interface RegraDeTaxaParaDesconto {
 
-	public double getTaxa(Item item);
+	public BigDecimal getTaxa(Item item);
 	public void setProximo(RegraDeTaxaParaDesconto regra);
 }
